@@ -1,40 +1,18 @@
-# Custom Bottom Sheet Dialog Example
+package com.prongbang.bottomsheetdialog
 
-### Result
-![screenshot gif](https://github.com/prongbang/images/blob/master/bottom-sheet-dialog.gif?raw=true)
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.FragmentManager
+import com.prongbang.sbs.SmartBottomSheetDialogFragment
+import kotlinx.android.synthetic.main.bottom_sheet_layout.*
 
-## Download
-
-> build.gradle
->
-```groovy
-allprojects {
-    repositories {
-        maven { url 'https://jitpack.io' }
-    }
-}
-```
-
-### AndroidX
-
-```groovy
-implementation 'com.github.prongbang:bottomsheetdialog:1.0.0'
-```
-
-## How to use
-
-> Add in bottom_sheet_layout.xml
-
-```xml
-app:layout_behavior="android.support.design.widget.BottomSheetBehavior"
-```
-
-> MyBottomSheetDialogFragment.kt
-
-```kotlin
 class MyBottomSheetDialogFragment(
 		supportFragmentManager: FragmentManager
 ) : SmartBottomSheetDialogFragment(supportFragmentManager) {
+
+	private var onClickCloseListener: (() -> Unit)? = null
 
 	override fun tagName(): String = MyBottomSheetDialogFragment::class.java.simpleName
 
@@ -50,7 +28,15 @@ class MyBottomSheetDialogFragment(
 
 	private fun initView() {
 		arguments?.let { args ->
-			// get value by key
+			context?.let {
+				ImageUtil.loadImageFromUrl(it, args.getString(AVATAR_URL_KEY), ivAvatar)
+			}
+			tvTitle.text = args.getString(TITLE_KEY)
+			tvSubTitle.text = args.getString(SUBTITLE_KEY)
+			ivClose.setOnClickListener {
+				dismiss()
+				onClickCloseListener?.invoke()
+			}
 		}
 	}
 
@@ -68,6 +54,26 @@ class MyBottomSheetDialogFragment(
 						}
 					}
 		}
+
+		fun setTitle(title: String): Builder {
+			argumentsBundle.putString(TITLE_KEY, title)
+			return this
+		}
+
+		fun setAvatar(url: String): Builder {
+			argumentsBundle.putString(AVATAR_URL_KEY, url)
+			return this
+		}
+
+		fun setSubtitle(subtitle: String): Builder {
+			argumentsBundle.putString(SUBTITLE_KEY, subtitle)
+			return this
+		}
+
+		fun setOnCloseButton(onListener: (() -> Unit)? = null): Builder {
+			onCloseListener = onListener
+			return this
+		}
 	}
 
 	companion object {
@@ -76,19 +82,3 @@ class MyBottomSheetDialogFragment(
 		private const val AVATAR_URL_KEY = "AVATAR_URL_KEY"
 	}
 }
-```
-
-> Use in Activity
-```kotlin
-MyBottomSheetDialogFragment.Builder(supportFragmentManager)
-    .setAvatar(avatarUrl)
-    .setTitle("เดฟไปวันๆ")
-    .setSubtitle("https://prongbang.github.io")
-    .setOnCloseButton {
-        Toast.makeText(this, "Bottom Sheet Dialog Closed", Toast.LENGTH_SHORT).show()
-    }
-    .build()
-    .show()
-```
-
-### Enjoy :V
